+++ services/harmony/app/util/umm-vis.ts
@@ -1,0 +1,11 @@
+export interface CmrUmmVisualization {
+  meta: {
+    // eslint-disable-next-line @typescript-eslint/naming-convention
+    'concept-id': string;
+    associations?: {
+      collections?: string[];
+      variables?: string[];
+    }
+  };
+  umm: object;
+}