+++ services/harmony/app/schemas/data-operation/CHANGELOG.md
@@ -1,3 +1,8 @@
+## [0.22.0] - 2025-05-29
+### Added
+- DataSource.visualizations - The UMM-Vis records associated with the collection
+- Variable.visualizations - The UMM-Vis records associated with a variable
+
 ## [0.21.0] - 2025-02-18
 ### Added
 - pixelSubset - true/false - true if pixel subset should be performed by the service.
@@ -12,7 +17,7 @@
 
 ## [0.18.0] - 2023-08-28
 ### Added
-- extendDimensions - The dimensions to extend (added inititally for TEMPO users)
+- extendDimensions - The dimensions to extend (added initially for TEMPO users)
 
 ## [0.17.0] - 2022-07-12
 ### Added